refactor(search): extract pure filter to lib/searchIndex.ts + unit tests
filterSearchIndex/SearchHit/SearchHitGroup don't depend on React or
Chakra, so live them in lib/ where unit tests can import them without
booting the full Chakra runtime (jsdom test env lacks structuredClone
which Chakra's preset needs). useGlobalSearchIndex re-exports the
public surface so existing consumers stay one import away.

Adds tests/unit/global-search-index.test.ts covering empty-query
launcher mode, multi-field matching, title>subtitle>keyword ranking,
AND semantics across words, case-insensitivity, and the canonical
group ordering.
